Uncomment fix for memory corruption
When appending an env value to an env, it was doing a full memcpy of
the original env val into the env's memory. However, this was problematic
because an env val contains a pointer to directly after its own memory
for the value it is actually holding on to. This was causing a pointer
to an old value somewhere on the heap to remain accessible even after
the memory had been freed.

In this particular case, the issue was exhibiting itself through an
assignment to something completely unrelated to an env val writing to
the location of the original pointer. This issue has been fixed by
re-assigning the pointer to the correct location in memory for the
env the val is being copied into.

However I just realized that this will still be a problem when the
env resizes due to too many vals being stored internally, so some
refactoring for how env vals store values internally is going to have
to be prioritized now to prevent memory corruption issues from
popping up all over the place.