Add initial capability for creating objects and handling messages
Currently an FST_Object is a full container for an environment and a
set of messages. In the future it is possible that an object will
contain a reference to a set of messages that are re-used between
objects instead so that message definition aren't being copied all
over the place whenever a new object is created.

Message handling is being tested by adding a println message to an int
object and asking for it to be handled.