Add classes that objects re-use for message definitions
Classes are stored at an interpreter level. There is currently no
support for deleting a class that has been created in an interpreter.
The interpreter is now passed to more locations in the code due to
classes being stored in it, so that code can potentially look up
classes if necessary.

A potential future optimization would be to special-case classes such
as those built in for primitives, so that lookup for them is the cost
of accessing the memory to them instead of having to loop over a list
of classes. For now the path is being left generalized so everything is
used the same to keep the interpreter simple during initial
implementation. This optimized path is likely to also have a function
that takes in an enum for the separately-stored "primitives" to allow
you to retrieve them generically.

Objects can have message added to themselves on an individual basis
still, their messages are not purely based on their class. There is no
way currently to create an inheritance hierarchy for classes, a single
object just gets access to one class. Likely in the future classes will
be given a parent hierarchy chain that mimics how Java or prototype
based OOP does. (this will be more prototype-based like Javascript, as
adding a message anywhere in the chain will cause all objects with that
class to immediately be able to respond to the new message)

Code has also been updated for all the new things that need to get set
in the various structs. New fields have been added and leaving those
unset would cause problems.