Fix using wrong pointer in array memset
Instead of using the beginning of the internally-allocated array
pointer, it was instead using the array pointer that was offset to the
start of where values would be stored. However, it was still doing a
memset on the entire length of the allocated memory, causing it to
overwrite bytes outside of what it allocated.