MASSIVE refactoring
- Rename tons of stuff to what I think is a better naming scheme. It's
inspired by how Python names its own functions.
- Entirely remove the concept of "values", everything should basically
just be an Object.
- Use the same ideology as Python to have "OOP" by having an Object
"head" that allows the entire struct to be cast to a pointer of the
base Object type even if it has more members. Implement a "Uint" type
using this. (might still change this to have "storage" or "property"
messages that can hold things to make Objects completely homogenous
even if they're holding a native int or something, but I still have
to figure out how to do that/if it's possible in a useful/performant
way)
- Simplify existing APIs a ton. There were so many overloaded functions
that did basically the same thing that were very unnecessary. There
is even more simplification coming as I probably refactor things like
the array to just hold things like Objects instead of trying to
hold anything inline to its own memory and whatnot which adds a
bunch of complexity. But the current refactoring is big enough
that I really need to just commit it.
- Update the test main code to make use of the new Uint type
for testing that all the mechanisms are working and get a sense
of the performance we currently have. Because we're doing
fewer allocations now with the more homogenous Object handling,
speed has increased. I still want to add Object pooling so that
we don't have to allocate as much by recycling Object definitions
which should make the speed be only what it takes to do the
lookup of message handlers.
- Still need to add in what I've learned about how ObjC does
its own message handling to increase performance.