Massively optimize object alloc/dealloc
We now use a memory pool instead of malloc/free
constantly. On my machine as of writing this code,
a full cycle of creating an int and adding it to
another int with a message is taking around 80
nanoseconds. It might take a bit less time than
this in actuality because it *seems* to take
around 17 nanoseconds to call
mach_absolute_time twice, so it could potentially
be around 63 nanoseconds total.

One thing to note: objects store the index
they are allocated at in the memory pool.
This is because the original code I wrote used
some pointer math with a division to figure
out the index an arbitrary object from the
pool was at without external code needing
to worry about internal indices. However, I
discovered that just by the virtue of having
a division that couldn't be optimized into
bit shift/and/or operations like the
existing divisions/modulus operations, it
was adding around 13 nanoseconds of
overhead to the memory pool free function.
Storing an index made the runtime drop
from 34 nanoseconds (or 17, compensating
for mach_absolute time) to 21 (or 4).
There's still room for the init path to
be optimized since it's still at 35/18
nanoseconds. But we're sitting a LOT better
than the like 300+ or 500+ we were at when
we had more allocations and used malloc/free
everywhere.