Optimization: keep track of indices to where holes in the memory pool are
This makes the "average" case for alloc/dealloc a bit
slower. But as a tradeoff, it ensures that even if
there are a lot of objects allocated (along with
random objects being freed among the pool) alloc
is still a constant-time operation. Previously it
could have taken N / 8 operations to find an open
hole, where N is the number of objects in the pool.
This would slow down massively as the number of
objects increased.