Attempt to make message handling safe for multithread
Also note that message finding/caching/etc. was moved
into the object code. This is because there was a
possibility that something would be returned from
finding a message handler in the class code and then
it could be deallocated after that point due to the
cleanup handler. Keeping all that code inlined to
the object code means that we can retrieve the
function pointer directly, which will never be
invalidated, then potentially let the cache be
freed.

It does cross concerns a bit having the code for
this in the object code, but if it means that we
can make the code fully parallel-safe then
I'd rather deal with that complexity.

Note that this isn't actually tested with
multiple threads yet, it's basically just
safe in theory. But the way it tries to be
safe is very similar to how the Objective-C
runtime does it, so it should basically be
ok with maybe a small bug or two that needs
to be hammered out as I write this code at
1:30am.